Rearranged directories for better layout.  Refactored vector<> into an ops base class and separate specializations for each vector type.  This allows more control over e.g. the external<> type to ensure copy assignment and constructing are not possible.

+++ tests/timing/cmatrix_algebra1.cpp
@@ -15,8 +15,8 @@ inline void timed1(
         )
 {
     matrix_d44 tmp, tmp2; 
-    for(int i = 0; i < n_iter; ++i) {
-        for(int row = 0; row < 4; ++row) {
+    for(size_t i = 0; i < n_iter; ++i) {
+        for(size_t row = 0; row < 4; ++row) {
             for(int col = 0; col < 4; ++col) {
 #if 0
                 m[row][col] = m[row][col] + m1[row][col] + m2[row][col]
